[FIX][Base] Fork can't catch error code
Problems:
- Fork found the error code from child process but it can't store
this code to _ECode. The reason is that the _ECode is stored locally
and usually it shouldn't be the Fork object that are used to check on
the test case.

Solution:
- Fork should share the same memory with the parameter _ECode and
we only free it when the creator is killed
- Add test case to verify 2 solution: using execvp only and use
another fork to keep track the child process that is using execvp